Migrated code related to the Activity feed to the  component.
